+++ src/libs/extensions/index.ts
@@ -1,2 +1,2 @@
-export * from "./masonry/masonry.extension"
-export * from "./swiper/index"
+export * as Masonry from "./masonry/masonry.extension"
+export * as Swiper from "./swiper/index"